chore: Error 클래스 및 메시지 내보내기 수정
