+++ backend/src/errors/message.ts
@@ -1,7 +1,7 @@
 export default {
   AUTH_WRONG: '아이디 또는 비밀번호가 일치하지 않습니다.',
-  BOOK_NOTFOUND: '일치하는 책이 없습니다.',
   AUTH_USERNAME_OVERLAP: '중복되는 아이디가 존재합니다.',
   AUTH_NICKNAME_OVERLAP: '중복되는 닉네임이 존재합니다.',
   ARTICLE_NOTFOUND: '일치하는 글이 없습니다.',
+  BOOK_NOTFOUND: '일치하는 책이 없습니다.',
 };
